Move parameter group calculation to a selector #59 [PR]
Parameter groups are basically just a different view on the parameter
state. This means that we can use a selector to calculate the parameter
groups. This also has the benefit of ensuring that parameter groups are
always in sync with parameters in the store and there is no interleaving
that would allow otherwise.
