+++ src/app/state/index.ts
@@ -1,7 +1,7 @@
 import {Type} from '@angular/core';
 import {FunctionalEffect} from '@ngrx/effects';
 import {ActionReducerMap, MetaReducer} from '@ngrx/store';
-import {extractParameterGroups, failLoadingCollectionParameters, loadCollectionParameters} from './parameters/effects/parameter.effects';
+import {failLoadingCollectionParameters, loadCollectionParameters} from './parameters/effects/parameter.effects';
 import {parameterFeature, parameterFeatureKey} from './parameters/reducers/parameter.reducer';
 import {ParameterState} from './parameters/states/parameter.state';
 
@@ -11,7 +11,5 @@ export interface State {
 export const reducers: ActionReducerMap<State> = {
   [parameterFeatureKey]: parameterFeature.reducer,
 };
-export const effects: (Type<unknown> | Record<string, FunctionalEffect>)[] = [
-  {loadCollectionParameters, failLoadingCollectionParameters, extractParameterGroups},
-];
+export const effects: (Type<unknown> | Record<string, FunctionalEffect>)[] = [{loadCollectionParameters, failLoadingCollectionParameters}];
 export const metaReducers: MetaReducer<State>[] = [];
