+++ src/app/state/parameters/actions/parameter.action.ts
@@ -1,12 +1,11 @@
 import {createActionGroup, props} from '@ngrx/store';
 import {errorProps} from '../../utils/error-props.util';
-import type {Parameter, ParameterGroup} from '../../../shared/models/parameter';
+import type {Parameter} from '../../../shared/models/parameter';
 
 export const parameterActions = createActionGroup({
   source: 'Parameters',
   events: {
     'Set loaded parameters': props<{parameters: Parameter[]}>(),
-    'Set parameter groups': props<{groups: ParameterGroup[]}>(),
     'Load parameter for collections': props<{collections: string[]}>(),
     'Set parameter loading error': errorProps(),
   },
