+++ src/app/state/parameters/effects/parameter.effects.ts
@@ -36,13 +36,3 @@ export const failLoadingCollectionParameters = createEffect(
   },
   {functional: true, dispatch: false},
 );
-
-export const extractParameterGroups = createEffect(
-  (actions$ = inject(Actions), parameterService = inject(ParameterService)) => {
-    return actions$.pipe(
-      ofType(parameterActions.setLoadedParameters),
-      map(({parameters}) => parameterActions.setParameterGroups({groups: parameterService.getParameterGroups(parameters)})),
-    );
-  },
-  {functional: true},
-);
