+++ src/app/state/parameters/states/parameter.state.ts
@@ -1,8 +1,7 @@
 import type {LoadingState} from '../../../shared/models/loading-state';
-import type {Parameter, ParameterGroup} from '../../../shared/models/parameter';
+import type {Parameter} from '../../../shared/models/parameter';
 
 export interface ParameterState {
   parameters: Parameter[];
   loadingState: LoadingState;
-  groups: ParameterGroup[];
 }
